DOC: Organize .md files in `doc/`
`doc/`'s markdown files were not organized, and had a lot of repeated information.
This commit (along with #42c8b97 and #38de72) should help address that.

There are still a number of gaps in doc coverage & some repeated information, but it should be easier to find specific information now.

Features:
- table of contents at ~/README.md and ~/doc/README.md
  - I put this into a tree that made sense to me, open to suggestions on how to re-organize. Thinking about breaking off parts of doc/usage/ to doc/internals, but it's not done yet.
  - Folders are Design (unchanged), Installing (package manager, cmake args, etc), and Usage (how to run, debug, and ensembles/indices).
  - Add note that the "real documentation" is in the works... see github.com/AetherModel/AetherDevelopment
     - I didn't want to touch this, but can migrate these README's there if it would be helpful.
- Add consistent markdown formatting. Did not change *too* much of the existing content. But re-organized it and fixed typos/style where necessary.

Spent a while debating whether to make this a GitHub Pages or add in to the RTD. Decided this would be easiest since it can be quickly adapted to either... But it's much less elegant that either would be. Navigation is particularly difficult, but I would imagine that people will start in the ~/README.md and then get linked to a specific page, then go back. Or will be hunting for a specific doc page which should be easier to find.

If it's not easier, I can re-organize or add more navigation options. I was afraid of deleting too much of the read-the-docs (since a lot of info there seems irrelevant), so didn't do anything there. But I can.

Note the code-of-conduct is still showing "problems" in a markdown linter, but I do not see it necessary to fix those. All other markdown files (outside of ~/.github/) show no errors.