BUG: ionGrid error prevents test success
This commit adds ionGrid to the default aether.json file (in share/run/aether.json).

The ionGrid specs set here hold no importance, but these three values need to be set to enable the tests to pass. Ideally, this is moved from the settings file to a function which reads the inputs (and either sets the default values or catches the error), but for now this works.